python-discord-bot/requirements.txt
# Discord API library
discord.py>=2.3.2

# HTTP client for API requests
httpx>=0.24.1

# Environment variable management
python-dotenv>=1.0.0

# Async utilities
asyncio>=3.4.3

# LLM APIs
openai>=1.3.0
anthropic>=0.5.0
mistralai>=0.0.7  # Mistral AI client

# Date and time handling
python-dateutil>=2.8.2

# For reading/writing JSON
pydantic>=2.0.0
json5>=0.9.14

# For uuid generation
uuid>=1.30

# For logging
loguru>=0.7.0

# For type annotations
typing-extensions>=4.7.1 